Make macOS modifiers able to detect releasing one
When both the left and right Command keys are pressed, releasing one of
them used to report a press, which was incorrect. This is now fixed by
noting how many copies of Command keys are currently pressed, and when
that count becomes 3 then that's a release. Same goes for other
modifiers. This solution is assuming that modifiers do not produce
native key-repeats, which is apparently the case on macOS.